==> notebooks/00_data_cleaning.ipynb <==
import pandas as pd

df = pd.read_csv("../data/raw/Headlines - Combined.csv")
df.head()

df.columns

df.columns=["snippet", "headline", "neutral", "punchy", "url"]
df.head()

df.to_csv("../data/raw/Headlines - Combined.csv", index=False)

import torch

if torch.cuda.is_available():
    print("✅ GPU is available and will be used.")
    print(f"Device Name: {torch.cuda.get_device_name(0)}")
    print(f"Total Memory: {torch.cuda.get_device_properties(0).total_memory / 1024**3:.2f} GB")
else:
    print("❌ GPU is NOT available. The training will run on the CPU.")
